changed Circle class

+++ src/main/java/core/basesyntax/Circle.java
@@ -10,7 +10,7 @@ public class Circle extends Figure {
 
     @Override
     public double calculateArea() {
-        return (int) (Math.PI * radius * radius);
+        return Math.PI * radius * radius;
     }
 
     @Override
